🧽 Clear The Trivial Pedantic Findings
`cargo clippy -- -W clippy::pedantic` reported 95 findings. It is not
part of CI, and the crate has never been clean against it. This commit
fixes the 49 that need only a mechanical change. No behaviour changes,
and no lint is allowed or silenced.

- `missing_errors_doc`: each `try_` constructor puts the sentence it
  already had under an `# Errors` heading. `FullGrid::try_new` no longer
  lists invalid dimensions, which it never returns.
- `redundant_closure_for_method_calls`: the metric constants name
  `Sq::manhattan`, `Sq::chebyshev`, and `Hex::distance` directly.
- Casts with a clean replacement. `Graph::slot` hands back the index's
  own `u32`, so no caller narrows a `usize`. `Graph` numbers a node with
  `u32::try_from`, which folds the conversion and the node limit into one
  failure. The engine turns a total into a `Cost` with `try_from`, which
  saturates where the cast truncated. The shape constructors widen a
  checked non-negative `i32` with `unsigned_abs`. `Grid::line` widens the
  distance, where it narrowed the cell count.
- Tests: `f32::midpoint`, digit separators, an elided lifetime, a `use`
  moved above the statements, one renamed variable, and backticks round a
  project name in doc comments.

46 findings remain, and each needs a decision: 37 numeric casts, most of
them deliberate float and lattice arithmetic; 7 `many_single_char_names`
in the lattice math; `manual_midpoint` on the offset halving in
`layout.rs`; and `missing_panics_doc` on `Path::destination`, whose
`expect` cannot fire.

+++ src/grid.rs
@@ -631,7 +631,7 @@ pub trait Grid: sealed::Sealed {
         // positions by scanning the board instead. This keeps the answer exact without making work
         // proportional to a coordinate a caller supplied.
         let distance = self.distance(lo, hi);
-        if distance > self.len() as u32 {
+        if distance as usize > self.len() {
             let mut cells: Vec<(u32, Idx)> = self
                 .indices()
                 .filter_map(|j| {
